students_attendance/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
qrcode==7.4.2
Pillow==10.0.0
pandas==2.1.0
plotly==5.16.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
